Build: respect $CC, $AR, $RM, $CFLAGS, $LDFLAGS
If the user has set a value for CC, AR, or RM, use that value.  GNU make
provides good defaults for all of these.

If the user has set CFLAGS, use that instead of -O2, but preserve the
other build options.

Append the local LDFLAGS to any value from the user.