build: migrate build system to autotools
Iniparser is mostly used as a subcomponent of other projects. Some
of them required it as shared library that is already present on
the target system, while others might want to incorporate it into
own package distribution and build/ldadd it to their final binary/
library statically.

Autotools suite supports a really nice feature called subpackages.
It allows project to have subpackage as it's dependency, and does
recursive ./configure / make handling by itself.

With proper autotools support, iniparser can be just dropped into
another project, add a few settings to configure.ac, lib/Makefile.am
and src/Makefile.am and we are ready to go.

For this to work (to be embedable statically into another project),
another way of building iniparser had to be implemented. It is
reachable by the switch --enable-subpackage. This causes main
library to be built as noinst (not-installing library), which in
turn causes libiniparser to produce only it's static version. This
is the desired result, as this causes parent project to link
iniparser statically into it's libs/binaries.