Updated test for when long int is not 8 byte
If sizeof(long int) != 8, the current tests will fail.  Update the test
to use limits.h to help know what LONG_MAX/MIN are.
This also implies that iniparser_getlongint() will vary in terms of
being able to accept different width integers, depending on
architecture.  The goal here was to fix the test, not change the API.