Add test to proof an inconsistency in iniparser concerning brackets
The section "12345]" is readable from dictionary directly after
iniparser_set().

The section is also written to file by iniparser_dump_ini() as "[12345]]"

But if the file is parsed by iniparser_load() it is read as "12345".

The reason is that iniparser_load() considers a section name to be
enclosed by "[" and "]". If it find a line enclosed by square brackets it
takes the characters enclosed by and excluding those brackets.

refs #147