Only find ruby executable as we are not linking to Ruby
While `find_package()` provides us with `RUBY_EXECUTABLE` it also tries to
find ruby libraries. This works perfectly on Linux, but it currently fails
to do so on Windows Gitlab runner.

```
find_program(RUBY_EXECUTABLE name ruby PATHS ENV PATH REQUIRED)
```

could be replaced by

```
set(RUBY_EXECUTABLE ruby)
```

Or we could just call `ruby` in the script. The advantage is that it will
fail on configure time if `RUBY_EXECUTABLE` is not found.