Replace custom `xstrdup()` by standard `strdup()`
Configuring with `-DCMAKE_UNITY_BUILD=ON` results in a build failure because
of multiple definitions of `xstrdup`.

Instead of disabling unity builds, or move `xstrdup()` into
[iniparser.h](src/iniparser.h) or giving each `xstrdup()` a unique name it
is replaced by `strdup()`

`xstrdup()` has been there since at least 2008 (7de83c39 and 88c50b7e).
With commit 1c91ade9 it was added to inner `src/dictionary.h` API.
But with commit 85023c97 this was undone and instead `_strdup()` was added
to `src/iniparser.c` to provide the same functionality.
With commit bcd92a48 `_strdup()` has been replaced by a copy of `xstrdup()`.

The reason given to introduce `xstrdup()` was:

```c
  This is a replacement for strdup(). This implementation is provided
  for systems that do not have it.
```

It is doubtful that such systems still exist today.

refs #181

Changelog: removed