Revert "Add test code to verify heap-Buffer-Overflow in dictionary_set()"
A C API cannot determine the length of a non-null-terminated const char
*val just from the pointer itself.

All we can do is to document the requirement that if you pass your
own strings to iniparser, make sure they are null-terminated.

So this test is reverted as it cannot be solved as long as the API takes
null-terminated C strings.

This reverts commit 3fe0828675b387e61d04b4a0cdb39454e7db3cc4.